Fix: 좋아요 조회 화면에서 사진 상세 화면으로 전환한 후에 좋아요 취소 및 다시 좋아요 기능 동작 시 crash 발생 문제 해결
- Realm에서 데이터가 삭제되어 존재하지 않는 상황이나, 존재하지 않는 데이터를 이용해 Realm에 새 데이터를 생성 시도함으로써 crash가 발생되었습니다.
- 삭제 전에 삭제할 데이터와 동일한 데이터를 백업해두고, Realm에 데이터 생성 시 백업한 데이터를 활용하는 방법으로 해결하였습니다.

+++ HelloMyPolaroid/PhotoDetail/ViewModel/PhotoDetailViewModel.swift
@@ -18,6 +18,7 @@ final class PhotoDetailViewModel {
     private var photo: Photo?
     
     private var likedPhoto: LikedPhoto?
+    private var backupLikedPhotoData: LikedPhoto?
     
     //MARK: - Inputs
     
@@ -188,7 +189,7 @@ final class PhotoDetailViewModel {
             }
         } else {
             
-            if let likedPhoto = self.likedPhoto {
+            if let likedPhoto = self.backupLikedPhotoData {
                 
                 self.repository.create(data: likedPhoto) { [weak self] result in
                     switch result {
@@ -207,6 +208,7 @@ final class PhotoDetailViewModel {
                             }
                         }
                         
+                        self?.likedPhoto = likedPhoto //데이터 다시 채워두기
                         self?.outputIsLiked.value = true
                         
                     case .failure(let error):
@@ -243,19 +245,24 @@ final class PhotoDetailViewModel {
         } else {
             if let likedPhoto = self.likedPhoto {
                 
-                self.repository.deleteItem(photoID: likedPhoto.photoID) { [weak self] result in
-                    switch result {
-                    case .success(let success):
-                        print(success)
-                        ImageFileManager.shared.removeImageFromDocument(filename: likedPhoto.photoID)
-                        ImageFileManager.shared.removeImageFromDocument(filename: likedPhoto.userProfileID)
-                        
-                        self?.outputIsLiked.value = false
-                        
-                    case .failure(let error):
-                        print(error)
-                        
-                        self?.outputIsLiked.value = true
+                self.backupLikedPhotoData = LikedPhoto(value: likedPhoto)
+                
+                if let backupData = self.backupLikedPhotoData {
+                    
+                    self.repository.deleteItem(photoID: likedPhoto.photoID) { [weak self] result in
+                        switch result {
+                        case .success(let success):
+                            print(success)
+                            ImageFileManager.shared.removeImageFromDocument(filename: backupData.photoID)
+                            ImageFileManager.shared.removeImageFromDocument(filename: backupData.userProfileID)
+                            
+                            self?.outputIsLiked.value = false
+                            
+                        case .failure(let error):
+                            print(error)
+                            
+                            self?.outputIsLiked.value = true
+                        }
                     }
                 }
             }
